+++ python-backend/app/api/mine/models/mines.py
@@ -3,7 +3,7 @@ import uuid
 
 from sqlalchemy.orm import validates
 from sqlalchemy.dialects.postgresql import UUID
-from .mixins import AuditMixin, Base
+from ...utils.models_mixins import AuditMixin, Base
 from app.extensions import db
