+++ python-backend/tests/mines/models/test_mineral_tenure_xref_model.py
@@ -1,5 +1,5 @@
 from tests.constants import TEST_TENURE_ID
-from app.mines.models.mines import MineralTenureXref
+from app.api.mine.models.mines import MineralTenureXref
 
 
 # MineralTenureXref Class Methods
