+++ python-backend/tests/person/models/test_person_model.py
@@ -1,5 +1,5 @@
 from tests.constants import TEST_FIRST_NAME, TEST_SURNAME, TEST_PERSON_GUID, TEST_MANAGER_GUID, TEST_MINE_GUID
-from app.mines.models.party import Party
+from app.api.party.models.party import Party
 
 
 # Person Model Class Methods
